Add first fully working skill with cloneJob
Had to basically refactor how the requests were being processed. Actually
they are handled as post requests to the endpoint (and i skipped
that part.

From here on, what's left is to write a propper wrapper in order to
ensure that the calls are being handled correctly (Every intent should
be a controller).